+++ src/main/java/duke/task/Todo.java
@@ -1,9 +1,11 @@
+package duke.task;
+
 /**
  * A task with no time constraint associated with.
  */
 public class Todo extends Task{
     /**
-     * Constructor for the Task class.
+     * Constructor for the duke.task.Task class.
      * @param description Description of the task at hand.
      */
     public Todo(String description) {
@@ -11,8 +13,8 @@ public class Todo extends Task{
     }
 
     /**
-     * Provides a String representation of the Todo.
-     * @return A String representation of the Todo.
+     * Provides a String representation of the duke.task.Todo.
+     * @return A String representation of the duke.task.Todo.
      */
     public String toString() {
         return "[T][" + this.getStatusIcon() + "] " + this.description;
